sipmlify test running project
