+++ src/CssInliner.Tests/Tocsoft.CssInliner.Tests.csproj
@@ -55,12 +55,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FileMockHttpMessageHandler.cs" />
-    <Compile Include="GeneralTests.cs" />
-    <Compile Include="HtmlAssert.cs" />
-    <Compile Include="ObjectExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestCase.cs" />
+    <Compile Include="TestRunner.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
