+++ src/Frameworks/net40/CssInliner/net40 - Tocsoft.CssInliner.csproj
@@ -90,12 +90,42 @@
     <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\..\..\CssInliner\HtmlProcessor.cs">
+      <Link>HtmlProcessor.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\InlinerConfig.cs">
+      <Link>InlinerConfig.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\IResourceLoader.cs">
+      <Link>IResourceLoader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\NodeExtentions.cs">
+      <Link>NodeExtentions.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\Processor.cs">
+      <Link>Processor.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\Properties\AssemblyInfo.cs">
+      <Link>Properties\AssemblyInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\ResourceLoaders\FileResourceLoader.cs">
+      <Link>ResourceLoaders\FileResourceLoader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\ResourceLoaders\HttpResourceLoader.cs">
+      <Link>ResourceLoaders\HttpResourceLoader.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\ScopedProperty.cs">
+      <Link>ScopedProperty.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\ScopePropertyRank.cs">
+      <Link>ScopePropertyRank.cs</Link>
+    </Compile>
+    <Compile Include="..\..\..\CssInliner\StringProcessor.cs">
+      <Link>StringProcessor.cs</Link>
+    </Compile>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <Target Name="BeforeBuild">
-    <ItemGroup>
-      <Compile Include="..\..\..\CssInliner\**\*.cs" />
-    </ItemGroup>
-  </Target>
   <Import Project="..\..\..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets" Condition="Exists('..\..\..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" />
   <Target Name="EnsureBclBuildImported" BeforeTargets="BeforeBuild" Condition="'$(BclBuildImported)' == ''">
     <Error Condition="!Exists('..\..\..\packages\Microsoft.Bcl.Build.1.0.14\tools\Microsoft.Bcl.Build.targets')" Text="This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=317567." HelpKeyword="BCLBUILD2001" />
